fix: refactor middleware
